Remove old kafka implementation
