+++ lib/logstash-logger/version.rb
@@ -1,3 +1,3 @@
 module LogStashLogger
-  VERSION = "0.19.3"
+  VERSION = "1.0.0"
 end
